Fix concurrency problem during the classification of instances
While the access to the rConnection is synchronized, thus preventing concurrency problems, we are using a global variable to setup the instance to classify. This means that if 2 threads try to classify in instance, it may happen that thread_1 setups it's instance, and in the meanwhile thread_2 setups other instance, and both threads end classifying the same instance.

This commit changes the R functions used to classify instances to receive a new parameter with the instance to be classified. The convertion of an instance to data frame was moved to be called inside the syncronized method used to evaluate the instance.