Force Travis CI to run with JDK 8 as otherwise the build is currently failing
Our CI builds were failing with a non-understandable fatal error on the JVM when forking to run the junit tests.

The builds passed locally, so we suspected from JVM changes on CI. Looking on Travis CI news it seems they bumped the environments automatically to Java 11, so it must be related: https://travis-ci.community/t/install-of-oracle-jdk-8-failing/3038/8

Forcing the build to run with JDK 8 (which is what we urrently support at Feedzai) seems to have fixed the problem.