openml-r: Provide a UT method to implement the abstract method to test classify
The abstract UT in AbstractProviderModelBaseTest changed the UT test to a method to be flexible for all providers to use their model and instance.

Create a UT for each provider (Generic-R and Caret) to implement the method.
Change back the return value from method #classify() of models that are used on generic-R to the same value as in two previous commit.